+++ phpunit.xml.dist
@@ -11,15 +11,12 @@
 >
     <testsuites>
         <testsuite name="silex-cors-provider Test Suite">
-            <directory>src/Test</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
     <filter>
         <whitelist>
             <directory>src</directory>
-            <exclude>
-                <directory>src/Test</directory>
-            </exclude>
         </whitelist>
     </filter>
 </phpunit>
