+++ tests/CorsEnableControllerCollectionTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace JDesrosiers\Silex\Provider\Test;
+namespace JDesrosiers\Silex\Provider\Tests;
 
 use JDesrosiers\Silex\Provider\CorsServiceProvider;
 use Silex\Application;
 use Symfony\Component\HttpKernel\Client;
 
-class CorsEnableControllerTest extends \PHPUnit_Framework_TestCase
+class CorsEnableControllerCollectionTest extends \PHPUnit_Framework_TestCase
 {
     protected $client;
 
@@ -16,10 +16,11 @@ class CorsEnableControllerTest extends \PHPUnit_Framework_TestCase
         $app["debug"] = true;
         $app->register(new CorsServiceProvider());
 
-        $controller = $app->get("/foo", function () {
+        $controllers = $app["controllers_factory"];
+        $controllers->get("/", function () {
             return "foo";
         });
-        $app["cors-enabled"]($controller);
+        $app->mount("/foo", $app["cors-enabled"]($controllers));
 
         $app->get("/bar", function () {
             return "bar";
@@ -30,7 +31,7 @@ class CorsEnableControllerTest extends \PHPUnit_Framework_TestCase
 
     public function testEnabledPreflight()
     {
-        $this->client->request("OPTIONS", "/foo");
+        $this->client->request("OPTIONS", "/foo/");
         $response = $this->client->getResponse();
 
         $this->assertTrue($response->isEmpty());
@@ -47,7 +48,7 @@ class CorsEnableControllerTest extends \PHPUnit_Framework_TestCase
 
     public function testEnabledController()
     {
-        $this->client->request("GET", "/foo");
+        $this->client->request("GET", "/foo/");
         $response = $this->client->getResponse();
 
         $this->assertTrue($response->isOk());
